// src/types/resources/MedicationAdministration.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDateTime } from '../simpleTypes/DateTime';
import { TPeriod } from '../partials/Period';
import { TMedicationAdministrationPerformer } from '../partials/MedicationAdministrationPerformer';
import { TAnnotation } from '../partials/Annotation';
import { TMedicationAdministrationDosage } from '../partials/MedicationAdministrationDosage';

export type TMedicationAdministration = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  instantiates?: TUri[];
  partOf?: TReference[];
  status: string;
  statusReason?: TCodeableConcept[];
  category?: TCodeableConcept;
  medicationCodeableConcept?: TCodeableConcept;
  medicationReference?: TReference;
  subject: TReference;
  context?: TReference;
  supportingInformation?: TReference[];
  effectiveDateTime?: TDateTime;
  effectivePeriod?: TPeriod;
  performer?: TMedicationAdministrationPerformer[];
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  request?: TReference;
  device?: TReference[];
  note?: TAnnotation[];
  dosage?: TMedicationAdministrationDosage;
  eventHistory?: TReference[];
};